backend/requirements.txt
fastapi
uvicorn
chromadb
pydantic<2
sentence-transformers
python-dotenv
passlib[bcrypt]
python-jose